refactor: 불분명한 예외명 수정
