+++ src/main/java/gift/exception/member/DuplicateEmailException.java
@@ -1,0 +1,25 @@
+package gift.exception.member;
+
+public class DuplicateEmailException extends RuntimeException {
+
+    public DuplicateEmailException() {
+        super();
+    }
+
+    public DuplicateEmailException(String message) {
+        super(message);
+    }
+
+    public DuplicateEmailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DuplicateEmailException(Throwable cause) {
+        super(cause);
+    }
+
+    protected DuplicateEmailException(String message, Throwable cause, boolean enableSuppression,
+        boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
