+++ src/main/java/gift/exception/member/MemberExceptionHandler.java
@@ -1,10 +1,7 @@
-package gift.exception;
+package gift.exception.member;
 
-import java.util.List;
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
+import gift.exception.ErrorResult;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -13,8 +10,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class MemberExceptionHandler {
 
     @ResponseStatus(HttpStatus.FORBIDDEN)
-    @ExceptionHandler(MemberException.class)
-    public ErrorResult joinExHandle(MemberException e) {
+    @ExceptionHandler(DuplicateEmailException.class)
+    public ErrorResult joinExHandle(DuplicateEmailException e) {
         return new ErrorResult("회원가입 에러", e.getMessage());
     }
